✨ [BACK] Calculation for offset and station

+++ tracker-backend/TrackerApi/Models/Coordinate.cs
@@ -1,3 +1,3 @@
 namespace TrackerApi.Models;
 
-public record Coordinate(double X, double Y);
+public record Coordinate(float X, float Y);
